feat: use tea binary from system if available (#53)
As a proposal solution to this issue: https://code.forgejo.org/actions/forgejo-release/issues/51

We use the system installed tea binary if available or download it as it was done before if not.
The location of the binary is stored in a new variable `$TEA_BIN` which replaces the need for the `$ BIN_DIR` location.

I tested it on Alpine and Debian.

Co-authored-by: Rayan Mestiri <rayan.mestiri@gmail.com>
Co-authored-by: Rayan MESTIRI <rayan.mestiri@NLGRSSL600K.ratpsmart.local>
Reviewed-on: https://code.forgejo.org/actions/forgejo-release/pulls/53
Reviewed-by: earl-warren <earl-warren@noreply.code.forgejo.org>
Co-authored-by: sheik <sheik@noreply.code.forgejo.org>
Co-committed-by: sheik <sheik@noreply.code.forgejo.org>