Update Tea 0.9.0 -> 0.10.1 (#85)
This is generally a good idea.
But it also fixes my issues with this action. For comparison, here is a run with [v2.7.2](https://codeberg.org/Mynacol/delegacy-rpz/actions/runs/59) and with [this branch](https://codeberg.org/Mynacol/delegacy-rpz/actions/runs/58).

This is a split-off from https://code.forgejo.org/actions/forgejo-release/pulls/82 to have small PRs.

Reviewed-on: https://code.forgejo.org/actions/forgejo-release/pulls/85
Reviewed-by: earl-warren <earl-warren@noreply.code.forgejo.org>
Co-authored-by: Mynacol <mynacol@noreply.code.forgejo.org>
Co-committed-by: Mynacol <mynacol@noreply.code.forgejo.org>